Add PowerlevelSyncer class
