+++ lib/bridge/IrcBridge.js
@@ -655,6 +655,9 @@ IrcBridge.prototype._onEvent = Promise.coroutine(function*(baseRequest, context)
             }
         }
     }
+    else if (event.type === "m.room.power_levels") {
+        this.ircHandler.powerlevelSyncer.onMatrixPowerlevelEvent(event);
+    }
     return undefined;
 });
