+++ lib/bridge/IrcHandler.js
@@ -1,27 +1,20 @@
 /*eslint no-invalid-this: 0 consistent-return: 0*/
-"use strict";
-const Promise = require("bluebird");
 
+const Promise = require("bluebird");
 const stats = require("../config/stats");
 const BridgeRequest = require("../models/BridgeRequest");
 const IrcRoom = require("../models/IrcRoom");
 const MatrixRoom = require("matrix-appservice-bridge").MatrixRoom;
 const MatrixUser = require("matrix-appservice-bridge").MatrixUser;
 const MatrixAction = require("../models/MatrixAction");
-
 const Queue = require("../util/Queue.js");
 const QueuePool = require("../util/QueuePool.js");
-
 const QuitDebouncer = require("./QuitDebouncer.js");
+const PowerlevelSyncer = require("./PowerlevelSyncer.js");
 
 const JOIN_DELAY_MS = 250;
 const JOIN_DELAY_CAP_MS = 30 * 60 * 1000; // 30 mins
-const MODES_TO_WATCH = [
-    "m", // We want to ensure we do not miss rooms that get unmoderated.
-    "k",
-    "i",
-    "s"
-];
+
 
 const NICK_USERID_CACHE_MAX = 512;
 const LEAVE_CONCURRENCY = 10;
@@ -85,6 +78,8 @@ function IrcHandler(ircBridge, config) {
     "force-off" - Disabled, cannot be enabled.
     */
     this._mentionMode = config.mapIrcMentionsToMatrix || "on";
+
+    this.powerlevelSyncer = new PowerlevelSyncer(ircBridge);
 
     this.getMetrics();
 }
@@ -838,89 +833,7 @@ IrcHandler.prototype.onMode = Promise.coroutine(function*(req, server, channel, 
         (enabled ? ("+" + mode) : ("-" + mode)),
         channel, by, arg
     );
-
-    const privateModes = ["k", "i", "s"];
-    if (privateModes.indexOf(mode) !== -1) {
-        yield this._onPrivateMode(req, server, channel, by, mode, enabled, arg);
-        return;
-    }
-
-    if (mode === "m") {
-        yield this._onModeratedChannelToggle(req, server, channel, by, enabled, arg);
-        return;
-    }
-
-    // Bridge usermodes to power levels
-    let modeToPower = server.getModePowerMap();
-    if (Object.keys(modeToPower).indexOf(mode) === -1) {
-        // Not an operator power mode
-        return;
-    }
-
-    const nick = arg;
-    const matrixRooms = yield this.ircBridge.getStore().getMatrixRoomsForChannel(server, channel);
-    if (matrixRooms.length === 0) {
-        req.log.info("No mapped matrix rooms for IRC channel %s", channel);
-        return;
-    }
-
-    // Work out what power levels to give
-    const userPowers = [];
-    if (modeToPower[mode] && enabled) { // only give this power if it's +, not -
-        userPowers.push(modeToPower[mode]);
-    }
-
-    // Try to also add in other modes for this client connection
-    const bridgedClient = this.ircBridge.getClientPool().getBridgedClientByNick(
-        server, nick
-    );
-    let userId = null;
-    if (bridgedClient) {
-        userId = bridgedClient.userId;
-        if (!bridgedClient.unsafeClient) {
-            req.log.info(`Bridged client for ${nick} has no IRC client.`);
-            return;
-        }
-        const chanData = bridgedClient.unsafeClient.chanData(channel);
-        if (!(chanData && chanData.users)) {
-            req.log.error(`No channel data for ${channel}`);
-            return;
-        }
-        const userPrefixes = chanData.users[nick];
-
-        userPrefixes.split('').forEach(
-            prefix => {
-                const m = bridgedClient.unsafeClient.modeForPrefix[prefix];
-                if (modeToPower[m] !== undefined) {
-                    userPowers.push(modeToPower[m]);
-                }
-            }
-        );
-    }
-    else {
-        // real IRC user, work out their user ID
-        userId = server.getUserIdFromNick(nick);
-    }
-
-    // By default, unset the user's power level. This will be treated
-    // as the users_default defined in the power levels (or 0 otherwise).
-    let level = undefined;
-    // Sort the userPowers for this user in descending order
-    // and grab the highest value at the start of the array.
-    if (userPowers.length > 0) {
-        level = userPowers.sort((a, b) => b - a)[0];
-    }
-
-    req.log.info(
-        `onMode: Mode ${mode} received for ${nick} - granting level of ${level} to ${userId}`
-    );
-
-    const promises = matrixRooms.map((room) => {
-        return this.ircBridge.getAppServiceBridge().getIntent()
-            .setPowerLevel(room.getId(), userId, level);
-    });
-
-    yield Promise.all(promises);
+    yield this.powerlevelSyncer.onMode(req, server, channel, by, mode, enabled, arg);
 });
 
 IrcHandler.prototype._onModeratedChannelToggle = Promise.coroutine(function*(req, server, channel,
@@ -948,72 +861,6 @@ IrcHandler.prototype._onModeratedChannelToggle = Promise.coroutine(function*(req
     }
 });
 
-IrcHandler.prototype._onPrivateMode = Promise.coroutine(function*(req, server, channel, by,
-                                                mode, enabled, arg) {
-    // 'k' = Channel requires 'keyword' to join.
-    // 'i' = Channel is invite-only.
-    // 's' = Channel is secret
-
-    // For k and i, we currently want to flip the join_rules to be
-    // 'invite' to prevent new people who are not in the room from
-    // joining.
-
-    // For s, we just want to control the room directory visibility
-    // accordingly. (+s = 'private', -s = 'public')
-
-    // TODO: Add support for specifying the correct 'keyword' and
-    // support for sending INVITEs for virtual IRC users.
-    let matrixRooms = yield this.ircBridge.getStore().getMatrixRoomsForChannel(server, channel);
-    if (matrixRooms.length === 0) {
-        req.log.info("No mapped matrix rooms for IRC channel %s", channel);
-        return;
-    }
-
-    if (mode === "s") {
-        if (!server.shouldPublishRooms()) {
-            req.log.info("Not syncing publicity: shouldPublishRooms is false");
-            return Promise.resolve();
-        }
-        const key = this.ircBridge.publicitySyncer.getIRCVisMapKey(server.getNetworkId(), channel);
-
-        matrixRooms.map((room) => {
-            this.ircBridge.getStore().setModeForRoom(room.getId(), "s", enabled);
-        });
-        // Update the visibility for all rooms connected to this channel
-        return this.ircBridge.publicitySyncer.updateVisibilityMap(
-            true, key, enabled
-        );
-    }
-    // "k" and "i"
-    matrixRooms.map((room) => {
-        this.ircBridge.getStore().setModeForRoom(room.getId(), mode, enabled);
-    });
-
-    var promises = matrixRooms.map((room) => {
-        switch (mode) {
-            case "k":
-            case "i":
-                req.log.info((enabled ? "Locking room %s" :
-                    "Reverting %s back to default join_rule"),
-                    room.getId()
-                );
-                if (enabled) {
-                    return this._setMatrixRoomAsInviteOnly(room, true);
-                }
-                // don't "unlock"; the room may have been invite
-                // only from the beginning.
-                enabled = server.getJoinRule() === "invite";
-                return this._setMatrixRoomAsInviteOnly(room, enabled);
-            default:
-                // Not reachable, but warn anyway in case of future additions
-                req.log.warn(`onMode: Unhandled channel mode ${mode}`);
-                return Promise.resolve();
-        }
-    });
-
-    yield Promise.all(promises);
-});
-
 /**
  * Called when channel mode information is received
  * @param {Request} req The metadata request
@@ -1024,38 +871,7 @@ IrcHandler.prototype._onPrivateMode = Promise.coroutine(function*(req, server, c
  */
 IrcHandler.prototype.onModeIs = Promise.coroutine(function*(req, server, channel, mode) {
     req.log.info(`onModeIs for ${channel} = ${mode}.`);
-
-    // Delegate to this.onMode
-    let promises = mode.split('').map(
-        (modeChar) => {
-            if (modeChar === '+') {
-                return Promise.resolve();
-            }
-            return this.onMode(req, server, channel, 'onModeIs function', modeChar, true);
-        }
-    );
-
-    // We cache modes per room, so extract the set of modes for all these rooms.
-    const roomModeMap = yield this.ircBridge.getStore().getModesForChannel(server, channel);
-    const oldModes = new Set();
-    Object.values(roomModeMap).forEach((roomMode) => {
-        roomMode.forEach((m) => {oldModes.add(m)});
-    });
-    req.log.debug(`Got cached mode for ${channel} ${[...oldModes]}`);
-
-    // For each cached mode we have for the room, that is no longer set: emit a disabled mode.
-    promises.concat([...oldModes].map((oldModeChar) => {
-        if (!MODES_TO_WATCH.includes(oldModeChar)) {
-            return Promise.resolve();
-        }
-        req.log.debug(`${server.domain} ${channel}: Checking if '${oldModeChar}' is still set.`);
-        if (!mode.includes(oldModeChar)) { // If the mode is no longer here.
-            req.log.debug(`${oldModeChar} has been unset, disabling.`);
-            return this.onMode(req, server, channel, 'onModeIs function', oldModeChar, false);
-        }
-    }));
-
-    yield Promise.all(promises);
+    yield this.powerlevelSyncer.onModeIs(req, server, channel, mode);
 });
 
 /**
@@ -1105,13 +921,6 @@ IrcHandler.prototype.onMetadata = Promise.coroutine(function*(req, client, msg, 
     yield this.ircBridge.sendMatrixAction(adminRoom, botUser, notice, req);
 });
 
-IrcHandler.prototype._setMatrixRoomAsInviteOnly = function(room, isInviteOnly) {
-    return this.ircBridge.getAppServiceBridge().getClientFactory().getClientAs().sendStateEvent(
-        room.getId(), "m.room.join_rules", {
-            join_rule: (isInviteOnly ? "invite" : "public")
-        }, ""
-    );
-};
 
 IrcHandler.prototype.invalidateCachingForUserId = function(userId) {
     if (this._mentionMode === "force-off") {
